Action code to trigger before an update to a Phase
Action code to trigger before an update to a Phase.

V 1.0 just prevents the user from setting a Phase to a certain set of statuses (to be parameterized) before the child inspection, if it exists, is closed.  We don't want AiM to get in a state where it's "checkmated" itself by letting a Phase get to a certain status before its inspection is closed.